requirements.txt
Django==3.2.4
djangorestframework==3.12.4
pillow==8.2.0
django-cors-headers==3.7.0
djangorestframework-jwt==1.11.0